Use IoC for MapLibre effect handlers and extract MapLibre to single component
The effect handlers can now be bound later to MapState. Before the effect handlers are available all the effects will just be saved. Then once they are available all the effects are handled by the newly set effect handlers. The effect handlers for MapLibre specifically are extraced to `map-libre-effect-handlers.ts`.